Rename directory and files. Add to documentation
