--- Stage III/Stage_I_factors.csv
@@ -1,19 +1,0 @@
-,APOPN,APOPDEP,APOPSNR,AGDPN,ATXPY,ASCHF,ABOOK,ACPIU,ACPIM,ARETS,AWAGE,ASCHCI,ASCHCL,ASCHEI,ASCHEL,AINTS,ADIVS,ACGNS,ASOCSEC,AUCOMP
-2009,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1
-2010,1.0084,0.9989,1.0216,1.0378,1.0283,1.2958,1.2501,1.0168,1.0341,1.0171,1.0228,1.045,0.8686,1.0608,0.8839,0.831,1.1228,1.4961,1.0454,1.4395
-2011,1.0161,0.9947,1.044,1.0762,1.0922,2.1268,1.3004,1.0485,1.0658,1.0347,1.061,1.0917,0.8488,1.1094,0.8301,0.7149,1.1906,1.5301,1.0735,1.1059
-2012,1.0239,0.9907,1.0894,1.121,1.1489,2.0366,1.448,1.0704,1.1046,1.0316,1.1041,1.154,0.8178,1.3034,0.7754,0.6654,1.593,2.451,1.1519,0.8527
-2013,1.0317,0.9882,1.1287,1.1652,1.172,2.3437,1.5082,1.0862,1.1318,1.0488,1.1346,1.1569,0.8652,1.3096,0.7916,0.5991,1.3151,1.9169,1.211,0.6216
-2014,1.0394,0.9863,1.1671,1.2032,1.2157,2.2,1.4838,1.1091,1.0913,1.0448,1.1889,1.2129,0.8982,1.4197,0.7905,0.5589,1.5582,2.6796,1.2582,0.3985
-2015,1.0476,0.9859,1.2071,1.2447,1.27,1.6873,1.438,1.1294,1.1028,1.058,1.2456,1.2702,0.9406,1.3759,0.7661,0.5633,1.6589,3.1191,1.3222,0.363
-2016,1.0561,0.9868,1.2473,1.285,1.3236,1.3352,1.4021,1.1552,1.1177,1.0747,1.3044,1.3226,0.9794,1.3415,0.747,0.5813,1.7004,3.3961,1.3757,0.2839
-2017,1.0647,0.988,1.2885,1.3387,1.3808,1.6592,1.4063,1.1809,1.1323,1.0932,1.3613,1.38,1.0219,1.3455,0.7492,0.606,1.7579,3.4295,1.4354,0.2738
-2018,1.0733,0.9896,1.3317,1.3929,1.4397,1.8817,1.42,1.2079,1.1478,1.1103,1.4167,1.434,1.0619,1.3587,0.7565,0.631,1.8163,3.463,1.5316,0.2939
-2019,1.0819,0.9912,1.3769,1.4439,1.4974,1.9972,1.4356,1.2361,1.1637,1.1253,1.4682,1.4818,1.0973,1.3735,0.7648,0.6683,1.8788,3.4582,1.6428,0.3303
-2020,1.0904,0.9921,1.4244,1.4957,1.5601,2.1296,1.443,1.2649,1.1802,1.1385,1.5213,1.5293,1.1325,1.3806,0.7688,0.7224,1.9371,3.5203,1.7601,0.3767
-2021,1.0989,0.9939,1.4712,1.5542,1.629,2.2732,1.4718,1.2943,1.197,1.152,1.58,1.5853,1.1739,1.4082,0.7841,0.7892,1.994,3.6158,1.8787,0.3985
-2022,1.1073,0.9969,1.5199,1.6161,1.7022,2.4197,1.5188,1.3245,1.2143,1.1649,1.6422,1.6463,1.2191,1.4531,0.8091,0.8535,2.0556,3.7305,2.0083,0.4167
-2023,1.1157,0.9997,1.5679,1.6811,1.7764,2.5944,1.585,1.3554,1.2319,1.1779,1.7071,1.7124,1.2681,1.5165,0.8444,0.9144,2.1234,3.8642,2.1494,0.4367
-2024,1.124,1.0022,1.6149,1.7488,1.8537,2.7972,1.6609,1.3871,1.25,1.1911,1.7743,1.7828,1.3202,1.5891,0.8849,0.9672,2.1983,4.0123,2.3004,0.4604
-2025,1.1322,1.0047,1.6637,1.8196,1.9331,2.9831,1.7485,1.4197,1.2686,1.2044,1.8442,1.8574,1.3755,1.6729,0.9315,1.0146,2.2799,4.1747,2.4567,0.4822
-2026,1.1403,1.0071,1.7098,1.893,2.0163,3.1127,1.8434,1.4531,1.2875,1.2179,1.9167,1.9353,1.4332,1.7637,0.9821,1.0622,2.3693,4.3419,2.6235,0.5031
