+++ stage3/pufadj_factors.csv
@@ -1,0 +1,20 @@
+INT2010,INT2011,INT2012,INT2013,INT2014,INT2015,INT2016,INT2017,INT2018,INT2019,INT2020,INT2021,INT2022,INT2023,INT2024,INT2025,INT2026
+0.894570657511,1.10672363651,0.828155525174,1.13897644483,0.925133833279,1.00974066398,0.997233067712,1.00020025711,0.993180991967,0.978129160696,1.03494427851,0.996653685345,0.984010833382,1.01502858594,1.00912206668,1.00057050568,1.00755127661
+0.277936915309,0.889333330304,0.914660166907,0.784182359895,0.921968806824,1.00422248817,0.991679928803,0.997368680599,0.981250571127,0.984092169449,0.990237894643,0.984089065364,0.984480850273,0.985335460664,0.986413028265,0.985726747261,0.987704766664
+0.905851867059,0.744717716945,0.853336496358,0.754171760532,0.853323973582,0.97310803284,0.968920774411,0.963548507444,0.964583038123,0.976217494533,0.962030413092,0.978113383754,0.986114958659,0.968950764435,0.978708191657,0.973218969426,0.975313011801
+0.865550027979,0.822829719314,0.860587814233,0.76018499586,0.766595101667,0.96721560407,0.964638004354,0.955590516209,0.963459458638,0.986961985687,0.951965021376,0.980210026159,0.999678273231,0.967568205456,0.979529491113,0.977867562495,0.981617237919
+0.862370124368,0.888355355771,0.737688316334,0.885769589089,0.834162259466,0.976221484037,0.972178922261,0.962158997023,0.970806796955,0.990935243425,0.962385099605,0.983013505354,0.99712739969,0.97957997814,0.984344140113,0.986102444116,0.987866706414
+0.939010546172,0.800847021222,0.827657759705,0.856216681324,0.866174994362,0.962705727651,0.967356703977,0.973242287677,0.974760715767,0.985401142393,0.981347554774,0.990676796501,0.990910462587,0.991273191883,0.990170214158,0.990957746618,1.00540188392
+0.854594552658,0.888039758719,0.819199293758,0.926420260632,0.984158261912,0.985406121339,0.983979905487,0.975108780722,0.979385506194,0.986022836117,0.99508948464,0.989696689744,0.986460422566,0.99713019261,1.00254157769,0.993195231355,1.00222567621
+0.841921125239,0.9235984934,0.79655237166,0.893968942552,0.858449655851,1.00619878485,0.991287990837,0.995907303522,0.980894657558,0.976652536361,1.02778910538,0.991723304633,0.999346974682,1.00248630218,1.0005675355,1.0044852269,0.999078532844
+0.801319283113,0.980486770542,0.832759302563,0.833220635914,0.828309014512,1.00685432684,0.995553529264,1.00527108467,0.990647008563,0.984818779403,1.03939770396,1.00295505973,0.977917503104,1.03292190103,1.01183092771,1.0089063489,1.00390542243
+0.832911751911,0.904283950902,0.873719451169,0.876497424882,1.02501023984,1.03338627906,1.00451070918,1.00430214894,0.991040898293,0.993313181425,1.02465416707,1.00374286944,1.00775386156,1.0153760181,1.00181765029,1.00212945398,0.993520621968
+0.920278986285,1.02959505892,0.888181974719,0.997938730096,0.847681686262,1.03780482171,1.00946305895,0.994569189736,1.00030055116,1.011212485,0.983812168605,1.00147372804,1.01052235142,0.98372023439,0.998010404057,0.995701385524,0.99333281483
+0.932635890516,1.00779637961,1.00023811743,0.983074986778,1.07633139864,1.01453670092,0.994072263012,0.99550614219,0.993261483622,1.00770209347,0.978543989005,0.999407064025,1.01088837326,0.986951696507,0.996580369769,0.99296627178,0.998577424229
+1.02741582794,1.06734222294,1.05528059309,1.00508265524,1.03664003313,1.02772001432,1.00161104447,1.00388325279,0.996850161176,0.999772872887,1.01073687533,0.9988951009,0.999379082098,1.00600980075,1.00316090496,1.00433728051,1.00249225677
+1.0364097442,1.04626981483,1.02483064366,1.03613142335,0.951169538521,1.01317874185,1.00988364386,1.01042676914,0.996984974053,0.986755418934,1.02335732647,0.992160703194,0.984445713327,1.01439592485,1.01025943385,1.01289298367,1.01064361289
+1.06437329288,1.04299793962,0.915856372804,1.18838687541,1.03164883753,0.959524456305,1.00528874276,1.00857972272,1.02909632393,1.00901288536,1.01904956792,1.00518671827,0.998857529857,1.02526770226,1.01237259591,1.02204281459,1.01358549246
+1.0367531046,1.10077470334,1.00296501146,1.11871454573,1.06051639002,0.93761270802,1.00809737128,1.01907257264,1.03056596077,1.03235035292,1.02611452117,1.01031097612,1.00255607971,1.03185365897,1.01038355305,1.01296657159,1.00895519969
+1.06848339321,1.09006709516,0.966196397791,1.18281801887,1.01735202653,0.970836935276,1.0114498475,1.03491043712,1.06167542629,1.01400575457,1.01270962371,1.01875183058,0.99836032652,1.02207422903,1.01136054379,1.02212901854,1.01404750649
+1.0185821615,1.15686454569,1.16743843535,1.02526237543,1.06416167641,0.965537518163,1.03687421759,1.04278549179,1.09482859975,1.06921417865,0.977565695758,1.07745190473,1.01400379613,0.993720787965,1.00141632202,1.00528003009,1.00628883564
+2.84172042524,1.02765479069,1.38060882956,1.17732163416,1.04248041776,0.899348501668,1.04540950687,1.01616895332,1.04890455969,1.06092766188,0.948792988014,1.01592134238,1.04031483574,0.975648620522,1.0018815789,1.01325579112,1.02196559012
