+++ stage3/stage3_targets.csv
@@ -1,0 +1,20 @@
+,2009,2010,2011,2012,2013,2014
+INT_0,10146461,7334566,6650566,5352850,5498824,4688264
+INT_1,1839413,1351202,1024889,890567,637316,541622
+INT_2,2954846,2297484,1667178,1295730,923365,770262
+INT_3,4220067,3037863,2436749,1873500,1336677,1002678
+INT_4,4353458,3139792,2610015,1734456,1426057,1147555
+INT_5,3796052,3026555,2257069,1692309,1360845,1125760
+INT_6,3739114,2758061,2196920,1650647,1434796,1340811
+INT_7,7073003,5135235,4202142,3099336,2570678,2096189
+INT_8,6945983,4816299,4143277,3213452,2470257,1900496
+INT_9,17789757,12864476,9762845,8207663,6436264,5955852
+INT_10,13371783,10693890,8804139,7258960,6541807,5013047
+INT_11,27383438,20329225,17545274,14743049,13814525,13390182
+INT_12,19821252,16260707,15037876,13959386,13316556,12725861
+INT_13,10082356,9259771,8308330,7993170,7678930,6814230
+INT_14,4666796,4358893,3997306,3891275,3708081,3825975
+INT_15,2851354,2600623,2568668,2709135,2382959,2582189
+INT_16,7793008,7554584,7227844,7552710,7043219,7098608
+INT_17,4502551,4738471,4515068,5421035,4452991,4547632
+INT_18,14670540,18053545,15155518,19250383,17614564,17327068
